e2e: add relaying tests
